Hack the kernel, get the API working -> Null Pointer Exception
So I'm now sure and certain that Rockchip's IOMMU code is unmaintained.

No fucking call works with this shit. Every thing fails in some
bizarre way because the IOMMU code must be updated every once in a blue
moon !

So the whole IOMMU mess bullshit needs to die RIGHT NOW. We'll use
PRIME and GEM and let the DRM code handle the whole copying itself,
when needed, and see how we can import stuff from the DRM to the
VPU itself when needed.

Can't we preallocate DMA memory for that and reuse the same buffer ?
It's not like you can't predict the biggest size you could ever need
during a decoding phase...

Signed-off-by: Myy <myy@miouyouyou.fr>