The DRM IOMMU has nothing to do with DRM
It's *impressive* how this *thing* has NOTHING to do with DRM drivers.

The only relationship it has with it is that it knows the provided FD
is from an exported DMA Buffer, and the driver author knows that this
DMA Buffer is associated with a DRM driver.

Basically, you could set up a sound driver, export a DMA Buffer with
anything in it, send it to the "VCodec IOMMU DRM" code and it would
work... if the code was working in the first place that is !

Anyway, you won't believe how getting a good editor, a good IDE,
a nice theme and a good font size can do for you and your health !

That said, I know have a good overview of the code flow, and what
these functions are for...

We're advancing... very slowly but we're advancing...

Signed-off-by: Myy <myy@miouyouyou.fr>